feat: backstage directory (#987)
This adds a backstage directory, only for use by plugins (not user-accessible). A part can now write to an export directory, which gets migrated unfiltered to the backstage directory during the stage process. That backstage directory can then be used as other parts.

Includes a change to the cargo-package plugin that makes it write to backstage instead of stage.